+++ src/main/kotlin/lotto/LottoApplication.kt
@@ -1,5 +1,9 @@
 package lotto
 
+import lotto.controller.LottoController
+import lotto.domain.LottoShop
+import lotto.domain.RandomLottoGenerator
+
 fun main() {
     LottoController(LottoShop(RandomLottoGenerator())).play()
 }
