+++ src/main/kotlin/lotto/view/OutputView.kt
@@ -1,4 +1,8 @@
-package lotto
+package lotto.view
+
+import lotto.domain.Lotto
+import lotto.domain.Rank
+import lotto.domain.Ranks
 
 class OutputView {
