+++ src/test/kotlin/lotto/RankTest.kt
@@ -4,6 +4,7 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.datatest.WithDataTestName
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
+import lotto.domain.Rank
 
 class RankTest : FunSpec({
